feat: Bufwipeout changes (#33)
The tricky part with this was that both of these should be true:

If the user exits a terminal with `^D` or `exit`, then the drawer should
close _and stay closed_ when going to a new tab. In this case the user
is saying "go away drawer".

If the user has a drawer with nvim-tree open and closes the tab, the
drawer should remain open (they didn't manually close the drawer,
`BufWipeout` happening on tabclose is a nvim-drawer thing, not a vim
thing).